+++ buildSrc/src/main/kotlin/util/KotlinTargetDetails.kt
@@ -3,7 +3,6 @@ package util
 import org.jetbrains.kotlin.konan.target.Family
 import org.jetbrains.kotlin.konan.target.KonanTarget
 
-
 enum class KotlinTargetDetails(
   val presetName: String,
   val hasCoroutines: Boolean,
@@ -18,16 +17,20 @@ enum class KotlinTargetDetails(
   IOS_ARM32("iosArm32", true),
   IOS_ARM64("iosArm64", true),
   IOS_X64("iosX64", true),
+  IOS_SIMULATOR_ARM64("iosSimulatorArm64", true),
 
   WATCHOS_X86("watchosX86", true),
   WATCHOS_X64("watchosX64", true),
   WATCHOS_ARM64("watchosArm64", true),
   WATCHOS_ARM32("watchosArm32", true),
+  WATCHOS_SIMULATOR_ARM64("watchosSimulatorArm64", true),
 
   TVOS_ARM64("tvosArm64", true),
   TVOS_X64("tvosX64", true),
+  TVOS_SIMULATOR_ARM64("tvosSimulatorArm64", true),
 
   MACOS_X64("macosX64", true),
+  MACOS_ARM64("macosArm64", true),
 
   LINUX_ARM32_HFP("linuxArm32Hfp", false),
   LINUX_MIPS32("linuxMips32", false),